requirements.txt
flask
pytesseract
pillow
spacy
dateparser
parsedatetime
